Increases unit-test coverage, does required refectoring, and updates libraries
Add basic tests for MainModel, and refactor to support this

clean up

Make the initial search explicit, and not a side-effect of init

Previously the init method in MainPresenter would trigger a search for the empty string, which is treated as "give me the first 50 results". I prefer to move it out of init, and have it being a more explicit call by the view.

Add some happy-path tests for MainPresenter

Added some unit tests for MainPresenter to cover the very happy parts. Some refactoring will be needed to improve the testability here. See the TODO in MainPresenter::search

Library updates

Use the latest version on maven central for mockito