Removes the backoff handling timer from the presenter
I am not happy with this code living in the presenter. After running some tests, it does not seem to make a noticeable difference on a high-end device, although testing on a lower-end device may show that it is still required. If the functionality is required again then it should probably live in the FolketsDatabase class,